Move allMedia from VAllResultsGrid component to media store
